[mod/#6] HaphapColors를 Color.kt로 이동
