+++ app/src/main/java/com/haphap/app/core/designsystem/theme/Theme.kt
@@ -6,31 +6,7 @@ import androidx.compose.runtime.CompositionLocalProvider
 import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.ReadOnlyComposable
 import androidx.compose.runtime.staticCompositionLocalOf
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.text.TextStyle
-
-@Immutable
-data class HaphapColors(
-    val primary500: Color,
-    val primary100: Color,
-    val sub100: Color,
-    val sub200: Color,
-    val sub300: Color,
-    val sub400: Color,
-    val yellow: Color,
-    val gray50: Color,
-    val gray100: Color,
-    val gray200: Color,
-    val gray300: Color,
-    val gray400: Color,
-    val gray500: Color,
-    val gray600: Color,
-    val gray700: Color,
-    val gray800: Color,
-    val gray900: Color,
-    val white: Color,
-    val black: Color,
-)
 
 @Immutable
 data class HaphapTypography(
@@ -58,28 +34,6 @@ data class HaphapTypography(
     val c_r_11: TextStyle,
     val c_sb_10: TextStyle,
     val c_r_10: TextStyle,
-)
-
-private val defaultHaphapColors = HaphapColors(
-    primary500 = Primary500,
-    primary100 = Primary100,
-    sub100 = Sub100,
-    sub200 = Sub200,
-    sub300 = Sub300,
-    sub400 = Sub400,
-    yellow = Yellow,
-    gray50 = Gray50,
-    gray100 = Gray100,
-    gray200 = Gray200,
-    gray300 = Gray300,
-    gray400 = Gray400,
-    gray500 = Gray500,
-    gray600 = Gray600,
-    gray700 = Gray700,
-    gray800 = Gray800,
-    gray900 = Gray900,
-    white = White,
-    black = Black,
 )
 
 private val defaultHaphapTypography = HaphapTypography(
